requirements.txt
pandas>=2.0.0
scikit-learn>=1.3.0
plotly>=5.15.0
streamlit>=1.28.0
fpdf2>=2.7.0
pyngrok>=7.0.0
matplotlib>=3.7.0
scipy>=1.11.0
